+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ## Unreleased
 ### Changed
+- Module Creation: Support screen and page.screen plugins (#145)
+- Fix imports in activity template (#145)
 - Lift up @Inject and @AssistedInject annotation (#141)
 
 ## [2.1.4] - 2025-10-14